+++ Api/Module/ConfigurationInterface.php
@@ -8,12 +8,14 @@ namespace Kiliba\Connector\Api\Module;
 interface ConfigurationInterface
 {
     /**
-     * @return string[]
-     */
-    public function setConfigValue();
-
-    /**
+     * Return Magento configuration for Kiliba
      * @return string[]
      */
     public function getConfigValue();
+
+    /**
+     * Refresh website token for Kiliba module API authentication
+     * @return string[]
+     */
+    public function refreshToken();
 }